+++ src/doupool/__init__.py
@@ -1,3 +1,3 @@
 """DouPool desktop application."""
 
-__version__ = "0.3.4.1"
+__version__ = "0.3.5"